move travel upgrade ws confirmation to a screen
