+++ src/pages/Travel/TravelUpgrade.tsx
@@ -1,22 +1,21 @@
 import type {StackScreenProps} from '@react-navigation/stack';
-import React, {useState} from 'react';
+import React, {useEffect, useState} from 'react';
 import HeaderWithBackButton from '@components/HeaderWithBackButton';
-import Modal from '@components/Modal';
 import ScreenWrapper from '@components/ScreenWrapper';
 import ScrollView from '@components/ScrollView';
-import type {WorkspaceConfirmationSubmitFunctionParams} from '@components/WorkspaceConfirmationForm';
-import WorkspaceConfirmationForm from '@components/WorkspaceConfirmationForm';
+import useCurrentUserPersonalDetails from '@hooks/useCurrentUserPersonalDetails';
 import useLocalize from '@hooks/useLocalize';
 import useNetwork from '@hooks/useNetwork';
 import useOnyx from '@hooks/useOnyx';
 import useThemeStyles from '@hooks/useThemeStyles';
 import Navigation from '@libs/Navigation/Navigation';
 import type {TravelNavigatorParamList} from '@libs/Navigation/types';
+import {getActivePolicies, isPaidGroupPolicy} from '@libs/PolicyUtils';
 import UpgradeConfirmation from '@pages/workspace/upgrade/UpgradeConfirmation';
 import UpgradeIntro from '@pages/workspace/upgrade/UpgradeIntro';
 import CONST from '@src/CONST';
-import {createDraftWorkspace, createWorkspace} from '@src/libs/actions/Policy/Policy';
 import ONYXKEYS from '@src/ONYXKEYS';
+import ROUTES from '@src/ROUTES';
 import type SCREENS from '@src/SCREENS';
 
 type TravelUpgradeProps = StackScreenProps<TravelNavigatorParamList, typeof SCREENS.TRAVEL.UPGRADE>;
@@ -26,28 +25,21 @@ function TravelUpgrade({route}: TravelUpgradeProps) {
     const feature = CONST.UPGRADE_FEATURE_INTRO_MAPPING.travel;
     const {translate} = useLocalize();
     const {isOffline} = useNetwork();
-    const [introSelected] = useOnyx(ONYXKEYS.NVP_INTRO_SELECTED, {canBeMissing: true});
+    const [policies] = useOnyx(ONYXKEYS.COLLECTION.POLICY, {canBeMissing: false});
+    const {login: currentUserLogin} = useCurrentUserPersonalDetails();
+    const groupPaidPolicies = getActivePolicies(policies, currentUserLogin).filter(isPaidGroupPolicy);
 
     const [isUpgraded, setIsUpgraded] = useState(false);
-    const [shouldShowConfirmation, setShouldShowConfirmation] = useState(false);
 
-    const onSubmit = (params: WorkspaceConfirmationSubmitFunctionParams) => {
-        createDraftWorkspace(introSelected, '', false, params.name, params.policyID, params.currency, params.avatarFile as File);
-        setShouldShowConfirmation(false);
+    useEffect(() => {
+        if (groupPaidPolicies.length < 1) {
+            return;
+        }
         setIsUpgraded(true);
-        createWorkspace({
-            policyOwnerEmail: '',
-            makeMeAdmin: false,
-            policyName: params.name,
-            policyID: params.policyID,
-            engagementChoice: undefined,
-            currency: params.currency,
-            file: params.avatarFile as File,
-        });
-    };
+    }, [groupPaidPolicies.length]);
 
-    const onClose = () => {
-        setShouldShowConfirmation(false);
+    const openWorkspaceConfirmation = () => {
+        Navigation.navigate(ROUTES.TRAVEL_WORKSPACE_CONFIRMATION);
     };
 
     return (
@@ -61,30 +53,6 @@ function TravelUpgrade({route}: TravelUpgradeProps) {
                 title={translate('common.upgrade')}
                 onBackButtonPress={() => Navigation.goBack(route.params.backTo)}
             />
-            <Modal
-                type={CONST.MODAL.MODAL_TYPE.RIGHT_DOCKED}
-                isVisible={shouldShowConfirmation}
-                onClose={onClose}
-                onModalHide={onClose}
-                onBackdropPress={() => {
-                    onClose();
-                    Navigation.dismissModal();
-                }}
-                enableEdgeToEdgeBottomSafeAreaPadding
-            >
-                <ScreenWrapper
-                    style={[styles.pb0]}
-                    includePaddingTop={false}
-                    enableEdgeToEdgeBottomSafeAreaPadding
-                    shouldKeyboardOffsetBottomSafeAreaPadding
-                    testID={TravelUpgrade.displayName}
-                >
-                    <WorkspaceConfirmationForm
-                        onSubmit={onSubmit}
-                        onBackButtonPress={onClose}
-                    />
-                </ScreenWrapper>
-            </Modal>
             <ScrollView contentContainerStyle={styles.flexGrow1}>
                 {isUpgraded ? (
                     <UpgradeConfirmation
@@ -95,7 +63,7 @@ function TravelUpgrade({route}: TravelUpgradeProps) {
                 ) : (
                     <UpgradeIntro
                         feature={feature}
-                        onUpgrade={() => setShouldShowConfirmation(true)}
+                        onUpgrade={openWorkspaceConfirmation}
                         buttonDisabled={isOffline}
                         loading={false}
                         isCategorizing
